Adds an explanation of Quantum Computing, now works on mobile.
This commit changes a couple of things, I added mobile support so that it looks decent on phones and tablets, as well I added a white paper to the bottom when you scroll down on what a Quantum Computer actually is. In addition, I modified the CSS to be more maintainable.